Refactor fake repository test into service tests
